% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.ses-package.R
\docType{package}
\name{aws.ses-package}
\alias{aws.ses}
\alias{aws.ses-package}
\title{aws.ses}
\description{
AWS SES Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Simple Email
Service (SES) REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

