% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_quota}
\alias{get_quota}
\title{Get Quota}
\usage{
get_quota(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{sesPOST}}.}
}
\description{
Get Quota
}
\examples{
\dontrun{
get_quota()
} 
}

