% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_statistics}
\alias{get_statistics}
\title{Get Statistics}
\usage{
get_statistics(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{sesPOST}}.}
}
\description{
Get Statistics
}
\examples{
\dontrun{
get_statistics()
} 
}

