% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{sesPOST}
\alias{sesPOST}
\title{SES HTTP Requests}
\usage{
sesPOST(query = list(), region = Sys.getenv("AWS_DEFAULT_REGION",
  "us-east-1"), key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), ...)
}
\arguments{
\item{query}{A list containing query string parameters}

\item{region}{A character string specifying an AWS region.}

\item{key}{A character string specifying an AWS Access Key ID.}

\item{secret}{A character string specifying an AWS Secret Access Key.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\description{
Low-level SES POST function
}

