\name{aws.ses-package}
\alias{aws.ses-package}
\alias{aws.ses}
\docType{package}
\title{AWS SES Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Simple Email Service (SES) REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
