% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{sesPOST}
\alias{sesPOST}
\title{SES HTTP Requests}
\usage{
sesPOST(query = list(), headers = list(), body = NULL,
  verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"), key = NULL,
  secret = NULL, session_token = NULL, ...)
}
\arguments{
\item{query}{A list containing query string parameters}

\item{headers}{A list of headers to pass to the HTTP request.}

\item{body}{A list of query-like parameters to be passed as a form-encoded message body.}

\item{verbose}{A logical indicating whether to be verbose. Default is given by \code{options("verbose")}.}

\item{region}{A character string specifying an AWS region. See \code{\link[aws.signature]{locate_credentials}}.}

\item{key}{A character string specifying an AWS Access Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{secret}{A character string specifying an AWS Secret Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{session_token}{Optionally, a character string specifying an AWS temporary Session Token to use in signing a request. See \code{\link[aws.signature]{locate_credentials}}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\description{
Low-level SES POST function
}
